# docs/swagger.yaml
definitions:
...
  forms.NavAddForm:
    properties:
      icon:
        type: string
      name:
        type: string
      open_type:
        type: integer
      parent_nav_id:
        type: integer
      url:
        type: string
    required:
    - name
    - url
    type: object
  forms.NavEditForm:
    properties:
      icon:
        type: string
      id:
        type: integer
      name:
        type: string
      open_type:
        type: integer
      parent_nav_id:
        type: integer
      url:
        type: string
    required:
    - id
    - name
    - url
    type: object
...
  utils.Result:
    properties:
      code:
        description: 状态码
        example: 100
        type: integer
      data:
        description: 数据
        type: object
      msg:
        description: 提示
        example: 密码错误
        type: string
    type: object
host: localhost:8088
info:
  contact: {}
  description: Aries 开源博客项目 API 接口文档
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Gin Swagger
  version: "1.0"
paths:
...
  /api/v1/all_links:
    get:
      consumes:
      - application/json
      responses:
        "100":
          description: Continue
          schema:
            $ref: '#/definitions/utils.Result'
        "104":
          schema:
            $ref: '#/definitions/utils.Result'
      summary: 获取所有友链
      tags:
      - 友链
...
  /api/v1/links:
    delete:
      consumes:
      - application/json
      parameters:
      - description: IDS
        in: query
        name: ids
        required: true
        type: string
      responses:
        "100":
          description: Continue
          schema:
            $ref: '#/definitions/utils.Result'
        "104":
          schema:
            $ref: '#/definitions/utils.Result'
      summary: 删除友链
      tags:
      - 友链
    get:
      consumes:
      - application/json
      responses:
        "100":
          description: Continue
          schema:
            $ref: '#/definitions/utils.Result'
        "104":
          schema:
            $ref: '#/definitions/utils.Result'
      summary: 分页获取友链
      tags:
      - 友链
    post:
      consumes:
      - application/json
      responses:
        "100":
          description: Continue
          schema:
            $ref: '#/definitions/utils.Result'
        "104":
          schema:
            $ref: '#/definitions/utils.Result'
      summary: 修改友链
      tags:
      - 友链
  /api/v1/links/{id}:
    delete:
      consumes:
      - application/json
      responses:
        "100":
          description: Continue
          schema:
            $ref: '#/definitions/utils.Result'
        "104":
          schema:
            $ref: '#/definitions/utils.Result'
      summary: 删除友链
      tags:
      - 友链
  /api/v1/navs:
    delete:
      consumes:
      - application/json
      parameters:
      - description: ids
        in: query
        name: ids
        required: true
        type: string
      responses:
        "100":
          description: Continue
          schema:
            $ref: '#/definitions/utils.Result'
        "104":
          schema:
            $ref: '#/definitions/utils.Result'
      summary: 批量删除菜单
      tags:
      - 菜单
    get:
      consumes:
      - application/json
      responses:
        "100":
          description: Continue
          schema:
            $ref: '#/definitions/utils.Result'
        "104":
          schema:
            $ref: '#/definitions/utils.Result'
      summary: 获取所有菜单
      tags:
      - 菜单
    post:
      consumes:
      - application/json
      parameters:
      - description: 添加菜单表单
        in: body
        name: addForm
        required: true
        schema:
          $ref: '#/definitions/forms.NavAddForm'
      responses:
        "100":
          description: Continue
          schema:
            $ref: '#/definitions/utils.Result'
        "104":
          schema:
            $ref: '#/definitions/utils.Result'
      summary: 添加菜单
      tags:
      - 菜单
    put:
      consumes:
      - application/json
      parameters:
      - description: 修改菜单表单
        in: body
        name: addForm
        required: true
        schema:
          $ref: '#/definitions/forms.NavEditForm'
      responses:
        "100":
          description: Continue
          schema:
            $ref: '#/definitions/utils.Result'
        "104":
          schema:
            $ref: '#/definitions/utils.Result'
      summary: 修改菜单
      tags:
      - 菜单
  /api/v1/navs/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      responses:
        "100":
          description: Continue
          schema:
            $ref: '#/definitions/utils.Result'
        "104":
          schema:
            $ref: '#/definitions/utils.Result'
      summary: 删除菜单
      tags:
      - 菜单
  /api/v1/navs/{type}/down/{order_id}:
    patch:
      consumes:
      - application/json
      parameters:
      - description: order_id
        in: path
        name: order_id
        required: true
        type: integer
      - description: type
        in: path
        name: type
        required: true
        type: string
      responses:
        "100":
          description: Continue
          schema:
            $ref: '#/definitions/utils.Result'
        "104":
          schema:
            $ref: '#/definitions/utils.Result'
      summary: 向移下动菜单
      tags:
      - 菜单
  /api/v1/navs/{type}/up/{order_id}:
    patch:
      consumes:
      - application/json
      parameters:
      - description: order_id
        in: path
        name: order_id
        required: true
        type: integer
      - description: type
        in: path
        name: type
        required: true
        type: string
      responses:
        "100":
          description: Continue
          schema:
            $ref: '#/definitions/utils.Result'
        "104":
          schema:
            $ref: '#/definitions/utils.Result'
      summary: 向上移动菜单
      tags:
      - 菜单
...
swagger: "2.0"